+++ dynamics/apis/__init__.py
@@ -9,9 +9,7 @@ from .attachments import Attachments
 from .employees import Employees
 from .locations import Locations
 from .dimensions import Dimensions
-from .dimension_values import DimensionValues
 from .bank_accounts import BankAccounts
-from .dimension_set_lines import DimensionSetLines
 
 __all__ = [
     'Companies',
@@ -25,7 +23,5 @@ __all__ = [
     'Employees',
     'Locations',
     'Dimensions',
-    'DimensionValues',
     'BankAccounts',
-    'DimensionSetLines'
 ]
