+++ dynamics/core/client.py
@@ -45,9 +45,7 @@ class Dynamics:
         self.employees = Employees()
         self.locations = Locations()
         self.dimensions = Dimensions()
-        self.dimension_values = DimensionValues()
         self.bank_accounts = BankAccounts()
-        self.dimension_set_lines = DimensionSetLines()
 
         # Get and set the access token
         access_token = self.__refresh_access_token()
@@ -74,9 +72,7 @@ class Dynamics:
             self.employees,
             self.locations,
             self.dimensions,
-            self.dimension_values,
             self.bank_accounts,
-            self.dimension_set_lines
         ]
 
         # Update access tokens in all API instances
@@ -103,9 +99,7 @@ class Dynamics:
             self.employees,
             self.locations,
             self.dimensions,
-            self.dimension_values,
             self.bank_accounts,
-            self.dimension_set_lines
         ]
 
         # Set batch URL for all API instances
@@ -134,9 +128,7 @@ class Dynamics:
             self.employees,
             self.locations,
             self.dimensions,
-            self.dimension_values,
             self.bank_accounts,
-            self.dimension_set_lines
         ]
 
         # Set base URL for all API instances
